--- src/main/common/form/models/mediation/MediationIndTelephoneNumber.ts
@@ -1,5 +1,0 @@
-import { YesNo } from '../../../../common/form/models/yesNo';
-export interface MediationIndTelephoneNumber {
-  option?: YesNo;
-  telephoneNumber?: string;
-}
