+++ src/main/common/models/mediation/freeMediation.ts
@@ -1,5 +1,5 @@
 import {IsDefined} from 'class-validator';
-import {VALID_YES_NO_OPTION} from '../../../../common/form/validationErrors/errorMessageConstants';
+import {VALID_YES_NO_OPTION} from '../../../common/form/validationErrors/errorMessageConstants';
 
 export class FreeMediation {
   @IsDefined({ message: VALID_YES_NO_OPTION })
