+++ src/main/common/models/mediation/mediation.ts
@@ -1,0 +1,12 @@
+import {MediationIndividualTelephoneNumber} from './mediationIndividualTelephoneNumber';
+import {FreeMediation} from './freeMediation';
+
+export class Mediation {
+  individualTelephone?: MediationIndividualTelephoneNumber;
+  mediationDisagreement?: FreeMediation;
+
+  constructor(individualTelephone?: MediationIndividualTelephoneNumber, mediationDisagreement?: FreeMediation) {
+    this.individualTelephone = individualTelephone;
+    this.mediationDisagreement = mediationDisagreement;
+  }
+}
