+++ src/main/common/models/mediation/mediationIndividualTelephoneNumber.ts
@@ -1,0 +1,5 @@
+import {YesNo} from '../../form/models/yesNo';
+export interface MediationIndividualTelephoneNumber {
+  option?: YesNo;
+  telephoneNumber?: string;
+}
