+++ src/main/views/features/mediation/mediation-disagreement.njk
@@ -9,7 +9,7 @@
 {% block content %}
 
   {% set yesHtml %}
-  <p class="govuk-body">{{ t('We’ll ask the claimant if they’ll try free mediation. If they say no, the claim will go to a hearing.') }}</p>
+  <p class="govuk-body">{{ t("We'll ask the claimant if they'll try free mediation. If they say no, the claim will go to a hearing.") }}</p>
   {% endset -%}
 
   <div class="govuk-width-container">
@@ -38,7 +38,7 @@
               <li>{{ t('mediation can be quicker and cheaper than going to court') }}</li>
               <li>{{ t('the mediator speaks to each side separately, you do not speak to the other party during mediation') }}</li>
               <li>{{ t('it gives you control over how your dispute is settled, which is not possible by going to court') }}</li>
-              <li>{{ t('it’s confidential and nothing said or done during mediation can be used in court') }}</li>
+              <li>{{ t("it's confidential and nothing said or done during mediation can be used in court") }}</li>
               <li>{{ t('mediation can avoid a ') }}
                 <a target="_blank" rel="noreferrer noopener" href="https://www.gov.uk/county-court-judgments-ccj-for-debt">{{  t('County Court Judgment (opens in a new tab)') }}</a>
                 {{ t(' being made against you') }}</li>
